requirements.txt
# TA-DSS: Post-Trade Position Monitoring System
# Dependency versions pinned for reproducibility
# Generated: 2026-02-27
# Python: 3.12.9

# =============================================================================
# BACKEND FRAMEWORK
# =============================================================================
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
starlette==0.35.1
pydantic-core==2.14.6

# =============================================================================
# DATABASE
# =============================================================================
sqlalchemy==2.0.25

# =============================================================================
# DATA ANALYSIS
# =============================================================================
pandas==2.3.3
pandas-ta==0.4.71b0
numpy==2.2.6
numba==0.61.2
llvmlite==0.44.0

# =============================================================================
# DATA SOURCES
# =============================================================================
ccxt==4.2.34
yfinance==0.2.36

# =============================================================================
# SCHEDULING
# =============================================================================
apscheduler==3.10.4

# =============================================================================
# NOTIFICATIONS
# =============================================================================
python-telegram-bot==22.6

# =============================================================================
# FRONTEND
# =============================================================================
streamlit==1.54.0
altair==6.0.0

# =============================================================================
# HTTP CLIENTS
# =============================================================================
httpx==0.28.1
httpcore==1.0.9
aiohttp==3.13.3
requests==2.32.5

# =============================================================================
# UTILITIES
# =============================================================================
python-dotenv==1.2.1
pyyaml==6.0.3
cryptography==46.0.5
certifi==2026.2.25
typing-extensions==4.15.0
annotated-types==0.7.0

# =============================================================================
# ASYNC SUPPORT
# =============================================================================
anyio==4.12.1
idna==3.11
h11==0.16.0

# =============================================================================
# TELEGRAM BOT DEPENDENCIES
# =============================================================================
tornado==6.5.4

# =============================================================================
# STREAMLIT DEPENDENCIES
# =============================================================================
altair==6.0.0
blinker==1.9.0
cachetools==6.2.6
gitpython==3.1.46
gitdb==4.0.12
packaging==26.0
pillow==12.1.1
protobuf==6.33.5
pyarrow==23.0.1
pydeck==0.9.1
tenacity==9.1.4
toml==0.10.2
tzdata==2025.3
jinja2==3.1.6
markupsafe==3.0.3
narwhals==2.17.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
referencing==0.37.0
rpds-py==0.30.0

# =============================================================================
# CCXT DEPENDENCIES
# =============================================================================
setuptools==82.0.0
cryptography==46.0.5
yarl==1.22.0
multidict==6.7.1
propcache==0.4.1
frozenlist==1.8.0
aiosignal==1.4.0
aiohappyeyeballs==2.6.1
attrs==25.4.0

# =============================================================================
# YFINANCE DEPENDENCIES
# =============================================================================
lxml==6.0.2
appdirs==1.4.4
pytz==2025.2
frozendict==2.4.7
peewee==4.0.0
beautifulsoup4==4.14.3
soupsieve==2.8.3
html5lib==1.1
webencodings==0.5.1
multitasking==0.0.12

# =============================================================================
# SCHEDULER DEPENDENCIES
# =============================================================================
six==1.17.0
tzlocal==5.3.1

# =============================================================================
# NOTIFICATION DEPENDENCIES
# =============================================================================
# (python-telegram-bot has no additional pinned deps beyond httpx)

# =============================================================================
# UTILITIES DEPENDENCIES
# =============================================================================
charset-normalizer==3.4.4
urllib3==2.6.3
click==8.3.1
httptools==0.7.1
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0

# =============================================================================
# DNS/ASYNC DEPENDENCIES (aiodns)
# =============================================================================
pycares==5.0.1
cffi==2.0.0
pycparser==3.0

# =============================================================================
# PRODUCTION DATABASE (OPTIONAL - UNCOMMENT FOR POSTGRESQL)
# =============================================================================
# psycopg2-binary==2.9.9
